requirements.txt
langchain==0.1.0
chromadb==0.4.22
sentence-transformers==2.2.2
python-dotenv==1.0.0
pypdf==4.0.1
tiktoken==0.5.2
openai==1.0.0
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.0
tqdm==4.66.1
pytest==7.4.2
python-magic==0.4.27
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
